python-examples/08-workflow-basics/pyproject.toml
[project]
name = "polos-workflow-basics"
version = "0.1.0"
description = "Workflow basics example for Polos SDK"
requires-python = ">=3.10"
dependencies = [
    "polos-sdk",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
]
